# openapi.json
{
  "components": {
    "schemas": {
...
      "AssignBody": {
        "additionalProperties": false,
        "properties": {
          "dimension_key": {
            "title": "Dimension Key",
            "type": "string"
          },
          "term_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Term Ids",
            "type": "array"
          }
        },
        "required": [
          "dimension_key"
        ],
        "title": "AssignBody",
        "type": "object"
      },
...
      "CreateTermInput": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "CreateTermInput",
        "type": "object"
      },
...
      "DimensionDTO": {
        "additionalProperties": false,
        "properties": {
          "dimension_key": {
            "title": "Dimension Key",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "max_items": {
            "title": "Max Items",
            "type": "integer"
          },
          "min_items": {
            "title": "Min Items",
            "type": "integer"
          },
          "public_visible": {
            "title": "Public Visible",
            "type": "boolean"
          },
          "selection_mode": {
            "title": "Selection Mode",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "dimension_key",
          "version",
          "display_name",
          "selection_mode",
          "min_items",
          "max_items",
          "public_visible"
        ],
        "title": "DimensionDTO",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TermDTO": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dimension_key": {
            "title": "Dimension Key",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "dimension_key",
          "name",
          "slug",
          "status"
        ],
        "title": "TermDTO",
        "type": "object"
      },
...
      "UpdateTermInput": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdateTermInput",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "aiya-cms (cms)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/taxonomy/dimensions": {
      "get": {
        "operationId": "list_dimensions_api_v1_admin_taxonomy_dimensions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DimensionDTO"
                  },
                  "title": "Response List Dimensions Api V1 Admin Taxonomy Dimensions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Dimensions",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      }
    },
    "/api/v1/admin/taxonomy/dimensions/{dimension_key}/terms": {
      "get": {
        "operationId": "list_terms_api_v1_admin_taxonomy_dimensions__dimension_key__terms_get",
        "parameters": [
          {
            "in": "path",
            "name": "dimension_key",
            "required": true,
            "schema": {
              "title": "Dimension Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TermDTO"
                  },
                  "title": "Response List Terms Api V1 Admin Taxonomy Dimensions  Dimension Key  Terms Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Terms",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      },
      "post": {
        "operationId": "create_term_api_v1_admin_taxonomy_dimensions__dimension_key__terms_post",
        "parameters": [
          {
            "in": "path",
            "name": "dimension_key",
            "required": true,
            "schema": {
              "title": "Dimension Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTermInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Term",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      }
    },
    "/api/v1/admin/taxonomy/targets/{target_type}/{target_id}/terms": {
      "delete": {
        "operationId": "remove_target_terms_api_v1_admin_taxonomy_targets__target_type___target_id__terms_delete",
        "parameters": [
          {
            "in": "path",
            "name": "target_type",
            "required": true,
            "schema": {
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Target Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Target Terms",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      },
      "put": {
        "operationId": "assign_terms_api_v1_admin_taxonomy_targets__target_type___target_id__terms_put",
        "parameters": [
          {
            "in": "path",
            "name": "target_type",
            "required": true,
            "schema": {
              "title": "Target Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Target Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Assign Terms",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      }
    },
    "/api/v1/admin/taxonomy/terms/{term_id}": {
      "patch": {
        "operationId": "update_term_api_v1_admin_taxonomy_terms__term_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTermInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Term",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      }
    },
    "/api/v1/admin/taxonomy/terms/{term_id}/archive": {
      "post": {
        "operationId": "archive_term_api_v1_admin_taxonomy_terms__term_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "term_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Term Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Term",
        "tags": [
          "admin",
          "admin-taxonomy"
        ]
      }
    },
...
  },
...
}